requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
colorama==0.4.4
cssselect2==0.5.0
Django==3.2.9
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-tempus-dominus==5.1.2.16
environ==1.0
future==0.18.2
html5lib==1.1
lxml==4.8.0
Pillow==9.0.1
PyPDF2==1.26.0
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.8
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.63.0
webencodings==0.5.1
xhtml2pdf==0.2.6
